refactor(init): reuse Wrangler JSONC config utilities
Wrangler TOML init support should not carry a separate copied JSONC editor module when the codebase already has the shared JSONC and Wrangler JSON path convention from the static assets work.

Use the shared JSONC utility shape, resolve JSON Wrangler config paths through the project helper, and collapse the TOML editor to the narrower source-preserving cases covered by init behavior tests.
